correção de bug na parte do CRUD do Usuario

+++ backend/src/main/java/br/com/pinguins/tcc/backend/controllers/UsuarioController.java
@@ -29,7 +29,7 @@ public class UsuarioController {
     }
 
     @GetMapping(value = "/id/{id}")
-    public ResponseEntity<UsuarioDTO> findById(@PathVariable("id") Long id) {
+    public ResponseEntity<UsuarioDTO> findById(@PathVariable("id") Integer id) {
         UsuarioDTO usuarioDTO = usuarioService.findById(id);
 
         return ResponseEntity.ok(usuarioDTO);
@@ -45,17 +45,16 @@ public class UsuarioController {
 
     @PutMapping(value = "/update/{id}", produces = MediaType.APPLICATION_JSON_VALUE,
             consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UsuarioDTO> update(@PathVariable("id") Long id, @RequestBody UsuarioDTO user) {
+    public ResponseEntity<UsuarioDTO> update(@PathVariable("id") Integer id, @RequestBody UsuarioDTO user) {
         UsuarioDTO usuarioDTO = usuarioService.updateById(id, user);
 
         return ResponseEntity.ok().body(usuarioDTO);
     }
 
     @DeleteMapping(value = "/delete/{id}")
-    public ResponseEntity<String> deleteById(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deleteById(@PathVariable("id") Integer id) {
         usuarioService.deleteById(id);
 
         return ResponseEntity.ok("Apagado com sucesso!");
     }
-
 }
